// apps/cms/src/shared/api/generated/logs/schemas/programSettlementPaymentItemResponse.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * JA Korea CMS Backend API — Logs subset
 * Filtered for CMS logs management Orval codegen.
 * OpenAPI spec version: v9
 */

export interface ProgramSettlementPaymentItemResponse {
  id?: number;
  itemType?: string;
  useYn?: boolean;
}